+++ packages/abjad-convert/src/DirectConverter.ts
@@ -5,5 +5,8 @@ export abstract class DirectConverter implements IConverter {
 	public readonly numberOfConnection = 1
 	public abstract readonly from: Abjad
 	public abstract readonly to: Abjad
+	public getConverterPath(): Abjad[] {
+		return [this.from, this.to]
+	}
 	public abstract convert(text: string): string
 }