+++ packages/abjad-convert/CHANGELOG.md
@@ -1,6 +1,11 @@
 # abjad-convert Changelog
 <!-- https://keepachangelog.com/en/1.0.0/ -->
 
+## [0.11.0] 2024-11-01
+### Added
+- Support for Nabataean
+- getConverterPath() function to get the path of the converter
+
 ## [0.10.0] 2024-08-31
 ### Added
 - Support for Runic